refactor[#321] : Chip 컴포넌트 파일 분리
- Chip에 해당하는 코드를 SelectChip 컴포넌트로서 다른 파일로 분리하였음
